requirements.txt
streamlit
pymupdf
python-docx
openai
tiktoken
groq
